fix: uncomment test, fix import, finish move
